requirements.txt
discord.py>=2.3.2
wavelink>=3.4.1
python-dotenv>=1.0.0
spotipy>=2.23.0
soundcloud-lib>=0.5.0
